Login, logout, signup views; permissions for CRUD views + unit tests

+++ webapp/views.py
@@ -4,6 +4,9 @@ from django import forms
 from django.template import RequestContext
 from django.shortcuts import render_to_response, redirect
 from django.http import HttpResponse, HttpResponseBadRequest, Http404
+from django.contrib.auth import authenticate, login, logout
+from django.contrib.auth.decorators import login_required
+from django.contrib.contenttypes.models import ContentType
 from django.core.context_processors import csrf
 from django.core.exceptions import ValidationError
 from django.db.models import Q
@@ -16,12 +19,51 @@ from patients.models import PatientProfile, SafetyNetRelationship, PrimaryContac
 from doctors.models import DoctorProfile
 from reminders.models import ReminderTime, Prescription, Message, SentReminder
 
+from guardian.shortcuts import assign_perm, get_objects_for_user
+from guardian.models import UserObjectPermission, GroupObjectPermission
 from lockdown.decorators import lockdown
+
 
 def landing_page(request):
 	return HttpResponse(content="STAY TUNED", content_type="text/plain")
 
-class NewPatientForm(forms.Form):
+
+class UserLoginForm(forms.Form):
+	primary_phone_number = USPhoneNumberField()
+	password = forms.CharField(widget=forms.PasswordInput())
+
+	def clean_primary_phone_number(self):
+		primary_phone_number = self.cleaned_data['primary_phone_number'].strip()
+		return convert_to_e164(primary_phone_number)
+
+
+class UserRegistrationForm(forms.Form):
+	full_name = forms.CharField(max_length=80) 
+	primary_phone_number = USPhoneNumberField()
+	password1 = forms.CharField(widget=forms.PasswordInput())
+	password2 = forms.CharField(widget=forms.PasswordInput())
+
+	def clean_full_name(self):
+		full_name = self.cleaned_data['full_name'].strip()
+		if len(full_name) == 0:
+			raise ValidationError('No full name provided')
+		return full_name
+
+	def clean_primary_phone_number(self):
+		primary_phone_number = self.cleaned_data['primary_phone_number'].strip()
+		return convert_to_e164(primary_phone_number)
+
+	def clean(self):
+		password1 = self.cleaned_data.get('password1')
+		password2 = self.cleaned_data.get('password2')
+
+		if password1 and password1 != password2:
+			raise forms.ValidationError("Passwords don't match")
+
+		return self.cleaned_data
+
+
+class CreatePatientForm(forms.Form):
     full_name = forms.CharField(max_length=80)
     primary_phone_number = USPhoneNumberField()
 
@@ -35,7 +77,8 @@ class NewPatientForm(forms.Form):
     	primary_phone_number = self.cleaned_data['primary_phone_number'].strip()
     	return convert_to_e164(primary_phone_number)
 
-class UpdatePatientForm(NewPatientForm):
+
+class UpdatePatientForm(CreatePatientForm):
 	p_id = forms.IntegerField()
 
 	def clean_p_id(self):
@@ -43,6 +86,7 @@ class UpdatePatientForm(NewPatientForm):
 		if not PatientProfile.objects.filter(id=p_id).exists():
 			raise ValidationError('Patient does not exist')
 		return p_id
+
 
 class DeletePatientForm(forms.Form):
 	p_id = forms.IntegerField()
@@ -52,6 +96,7 @@ class DeletePatientForm(forms.Form):
 		if not PatientProfile.objects.filter(id=p_id).exists():
 			raise ValidationError('Patient does not exist')
 		return p_id
+
 
 class NewReminderForm(forms.Form):
 	p_id = forms.IntegerField()
@@ -91,6 +136,7 @@ class NewReminderForm(forms.Form):
 		)
 		self.cleaned_data['active_days_of_week'] = active_days_of_week
 		return self.cleaned_data
+
 
 class DeleteReminderForm(forms.Form):
 	p_id = forms.IntegerField()
@@ -133,49 +179,127 @@ class DeleteReminderForm(forms.Form):
 			self.cleaned_data['all_deleted'] = True
 		return self.cleaned_data
 
-def login(request):
-	pass
 
-# @lockdown()
-def fishfood(request):
+def user_registration(request):
 	c = RequestContext(request)
-	c['patient_search_results_list'] = PatientProfile.objects.filter(
-		Q(status=PatientProfile.ACTIVE) | Q(status=PatientProfile.NEW)).order_by('full_name')
+	if request.method == 'GET':
+		return render_to_response('fishfood/user_registration.html', c)
 
-	return render_to_response('fishfood/fishfood.html', c)
-
-# @lockdown()
-# create new patient
-def create_patient(request, *args, **kwargs):
-	# validate form
-	if request.POST:
-		form = NewPatientForm(request.POST)
+	if request.method == 'POST':
+		form = UserRegistrationForm(request.POST)
 		if form.is_valid():
 			full_name = form.cleaned_data['full_name']
-			name_tokens = full_name.split()
-			first_name = name_tokens[0].strip()
-			last_name = "".join(name_tokens[1:]).strip()
 			primary_phone_number = form.cleaned_data['primary_phone_number']
-			# create new patient
-			try: 
+			password = form.cleaned_data['password1']
+			try: # only patient registration for now
 				(patient, created) = PatientProfile.objects.get_or_create(
-					first_name=first_name, last_name=last_name, 
+					full_name=full_name,
 					primary_phone_number=primary_phone_number,
 				)
 			except ValidationError:
 				pass
 			else:
+				if not created:
+					return HttpResponse('This patient already exists.')
+				else:
+					assign_perm('manage_patient_profile', patient, patient)
+
+					patient.set_password(password)
+					patient.is_claimed = True
+					patient.save()
+					patient = authenticate(phone_number=primary_phone_number, password=password)
+					login(request, patient)
+					return redirect('/fishfood/')
+	return HttpResponseBadRequest('Something went wrong.')
+
+
+def user_login(request):
+	c = RequestContext(request)
+	if request.method == 'GET':
+		return render_to_response('fishfood/user_login.html', c)
+
+	if request.method == 'POST':
+		form = UserLoginForm(request.POST)
+		if form.is_valid():
+			primary_phone_number = form.cleaned_data['primary_phone_number']
+			password = form.cleaned_data['password']
+			user = authenticate(phone_number=primary_phone_number, password=password)
+			if user is not None:
+				if user.is_active:
+					login(request, user)
+					return redirect('/fishfood/')
+				else:
+					HttpResponseBadRequest('This user account is disabled.')
+		else:
+			HttpResponseBadRequest('Login is invalid.')
+	return HttpResponseBadRequest('Something went wrong.')
+
+
+@login_required
+def user_logout(request):
+	c = RequestContext(request)
+	if request.method == 'POST':
+		logout(request)
+		return redirect('/fishfood/')
+
+	return HttpResponseBadRequest('')
+
+
+# @lockdown()
+@login_required
+def fishfood(request):
+	if request.method == 'GET':
+		c = RequestContext(request)
+		user = request.user
+
+		results = get_objects_for_user(user, 'patients.manage_patient_profile')
+		results = results.filter(
+			Q(status=PatientProfile.ACTIVE) | Q(status=PatientProfile.NEW)).order_by('full_name')
+
+		c['patient_search_results_list'] = results
+		return render_to_response('fishfood/fishfood.html', c)
+	return HttpResponseBadRequest('Something went wrong.')
+
+# @lockdown()
+# create new patient
+@login_required
+def create_patient(request, *args, **kwargs):
+	# validate form
+	if request.POST:
+		form = CreatePatientForm(request.POST)
+		if form.is_valid():
+			full_name = form.cleaned_data['full_name']
+			primary_phone_number = form.cleaned_data['primary_phone_number']
+			# create new patient
+			try: 
+				(patient, created) = PatientProfile.objects.get_or_create(
+					full_name=full_name,
+					primary_phone_number=primary_phone_number,
+				)
+			except ValidationError:
+				pass
+			else:
+				# add permissions
+				if not patient.num_caregivers > 0:
+					assign_perm('manage_patient_profile', request.user, patient)
+				else:
+					# TODO(minqi): send request to become caregiver
+					return HttpResponseBadRequest('This user is already under management.')
+
 				c = RequestContext(request)
 				c['patient'] = patient
 				if created or patient.status == PatientProfile.QUIT:
 					patient.status = PatientProfile.NEW
+					patient.num_caregivers += 1
 					patient.save()
 					ReminderTime.objects.create_welcome_notification(to=patient)
-				return render_to_response('fishfood/patient_view.html', c)
-			return HttpResponseBadRequest("This patient already exists.")
+				return render_to_response('fishfood/patient_view.html', c) # change to redirect
+			return HttpResponseBadRequest("This user already exists.")
 	return HttpResponseBadRequest("Something went wrong.")
 
+
 # @lockdown()
+@login_required
 def retrieve_patient(request, *args, **kwargs):
 	if request.GET:
 		c = RequestContext(request)
@@ -189,6 +313,9 @@ def retrieve_patient(request, *args, **kwargs):
 			except PatientProfile.DoesNotExist:
 				pass
 			else:
+				if not request.user.has_perm('patients.manage_patient_profile', patient):
+					return HttpResponseBadRequest("You don't have access to this user's profile")
+
 				c['patient'] = patient
 				# get reminders
 				reminders = ReminderTime.objects.filter(
@@ -205,21 +332,27 @@ def retrieve_patient(request, *args, **kwargs):
 					reminder_groups.append(drug_group)
 				c['reminder_groups'] = tuple(reminder_groups)
 				return render_to_response('fishfood/patient_view.html', c)
-		return HttpResponseBadRequest("Patient does not exist.")
+		return HttpResponseBadRequest("This user already exists.")
 	return HttpResponseBadRequest("Something went wrong.")
 
+
 # @lockdown()
+@login_required
 def update_patient(request, *args, **kwargs):
 	if request.POST:
 		form = UpdatePatientForm(request.POST)
 		if form.is_valid():
+			p_id = form.cleaned_data['p_id']
+			patient = PatientProfile.objects.get(id=p_id)
+
+			if not request.user.has_perm('patients.manage_patient_profile', patient):
+				return HttpResponseBadRequest("You don't have access to this user's profile")
+
 			full_name = form.cleaned_data['full_name']
 			name_tokens = full_name.split()
 			first_name = name_tokens[0].strip()
 			last_name = "".join(name_tokens[1:]).strip()
 			primary_phone_number = form.cleaned_data['primary_phone_number']
-			p_id = form.cleaned_data['p_id']
-			patient = PatientProfile.objects.get(id=p_id)
 			patient.first_name = first_name
 			patient.last_name = last_name
 			patient.full_name = full_name
@@ -233,34 +366,61 @@ def update_patient(request, *args, **kwargs):
 			return HttpResponse(json.dumps(result), content_type='application/json')
 	return HttpResponseBadRequest('Something went wrong')
 
+
 # @lockdown()
+@login_required
 def delete_patient(request, *args, **kwargs):
 	if request.POST:
 		form = DeletePatientForm(request.POST)
 		if form.is_valid():
 			p_id = form.cleaned_data['p_id']
 			patient = PatientProfile.objects.get(id=p_id)
-			patient.status = PatientProfile.QUIT
-			patient.save()
-			# Delete patient's Prescriptions, Notifications
+
+			if not request.user.has_perm('patients.manage_patient_profile', patient):
+				return HttpResponseBadRequest("You don't have access to this user's profile")
+
+			# patient loses a caregiver
+			patient.num_caregivers -= 1
+
+			# if patient is request user, delete the patient 
+			if request.user == patient:
+				patient.num_caregivers = 0
+
+			# If patient has no more caregivers, delete patient's Prescriptions, Notifications
 			# and also SafetyNetRelationship, PrimaryContactRelationship objects for which
 			# the patient is the source_patient
 			# Note we keep outstanding Messages and SentReminders
-			Prescription.objects.filter(patient=patient).delete()
-			ReminderTime.objects.filter(to=patient).delete()
-			SafetyNetRelationship.objects.filter(source_patient=patient).delete()
-			PrimaryContactRelationship.objects.filter(source_patient=patient).delete()
+			if patient.num_caregivers == 0:
+				Prescription.objects.filter(patient=patient).delete()
+				ReminderTime.objects.filter(to=patient).delete()
+				SafetyNetRelationship.objects.filter(source_patient=patient).delete()
+				PrimaryContactRelationship.objects.filter(source_patient=patient).delete()
+				# delete permissions
+				filters = Q(content_type=ContentType.objects.get_for_model(patient), 
+			    	object_pk=patient.pk)
+				UserObjectPermission.objects.filter(filters).delete()
+				GroupObjectPermission.objects.filter(filters).delete()
+				patient.quit()
+
+			patient.save()
+
 			return redirect('/fishfood/')
+
 	return HttpResponseBadRequest('Something went wrong')
 
+
 # @lockdown()
+@login_required
 def patient_search_results(request, *args, **kwargs):
 	if request.GET:
 		q = request.GET['q']
-		results = PatientProfile.objects.filter(
+
+		results = get_objects_for_user(request.user, 'patients.manage_patient_profile')
+		results = results.filter( 
 			Q(full_name__icontains=q) &
 			(Q(status=PatientProfile.ACTIVE) | Q(status=PatientProfile.NEW))
 		).order_by('full_name')
+
 		return render_to_response(
 			'fishfood/patient_search_results_list.html', 
 			{'patient_search_results_list':results}
@@ -268,12 +428,9 @@ def patient_search_results(request, *args, **kwargs):
 	else:
 		return HttpResponseBadRequest("Something went wrong")
 
-# @lockdown()
-def patient_reminder_list(request, *args, **kwargs):
-	reminder_id = request.GET['id']
-	(reminder, exists) = ReminderTime.objects.get(id=reminder_id)
 
 # @lockdown()
+@login_required
 def create_reminder(request, *args, **kwargs):
 	# update if reminder exists, else create it'
 	if request.POST:
@@ -285,22 +442,22 @@ def create_reminder(request, *args, **kwargs):
 
 			p_id = form.cleaned_data['p_id']
 			patient = PatientProfile.objects.get(id=p_id)
+
+			if not request.user.has_perm('patients.manage_patient_profile', patient):
+				return HttpResponseBadRequest("You don't have access to this user's profile")
 			
 			drug_name = form.cleaned_data['drug_name']
 			drug = Drug.objects.get_or_create(name=drug_name)[0]
 
-			dr_smartdose = DoctorProfile.objects.get_or_create(
-				first_name="Smartdose", last_name="", 
-				primary_phone_number="+18569067308", 
-				birthday=datetime.date(2014, 1, 28))[0]
 			(prescription, prescription_created) = Prescription.objects.get_or_create(
-				prescriber=dr_smartdose, patient=patient, drug=drug)
+				prescriber=request.user, patient=patient, drug=drug)
 
 			reminder_time = form.cleaned_data['reminder_time']
 			existing_reminders = ReminderTime.objects.filter(
 				to=patient, prescription__drug__name__iexact=drug_name)
 
 			# check if it's a daily reminder
+			new_reminders = []
 			is_daily_reminder = False not in active_days_of_week
 			med_reminder = None
 			if is_daily_reminder:
@@ -316,6 +473,7 @@ def create_reminder(request, *args, **kwargs):
 					send_time = send_datetime, 
 					repeat=ReminderTime.DAILY, 
 					prescription=prescription)[0]
+				new_reminders.append(med_reminder)
 				med_reminder.update_to_next_send_time()
 				med_reminder.day_of_week = 8
 				med_reminder.save()
@@ -346,6 +504,7 @@ def create_reminder(request, *args, **kwargs):
 								send_time = send_datetime, 
 								repeat=ReminderTime.WEEKLY, 
 								prescription=prescription)[0]
+							new_reminders.append(med_reminder)
 							med_reminder.day_of_week = idx + 1
 							med_reminder.save()
 			# create a refill reminder for the patient if prescription is not filled
@@ -356,29 +515,41 @@ def create_reminder(request, *args, **kwargs):
 					reminder_type=ReminderTime.REFILL, 
 					repeat=ReminderTime.DAILY, 
 					prescription=prescription)[0]
+				new_reminders.append(refill_reminder)
 			elif not send_refill_reminder and prescription_created:
 				prescription.filled = True
 				prescription.save()
+			for r in new_reminders:
+				assign_perm('view_notification', request.user, r)
+				assign_perm('change_notification', request.user, r)
 			return HttpResponse('')
 	return HttpResponseBadRequest()
 
+
 # @lockdown()
+@login_required
 def update_reminder(request, *args, **kwargs):
 	pass
 
+
 # @lockdown()
+@login_required
 def delete_reminder(request, *args, **kwargs):
 	if request.POST:
 		form = DeleteReminderForm(request.POST)
 		if form.is_valid():
 			p_id = form.cleaned_data['p_id']
 			patient = patient = PatientProfile.objects.get(id=p_id)
+			
+			if not request.user.has_perm('patients.manage_patient_profile', patient):
+				return HttpResponseBadRequest("You don't have access to this user's profile")
+
 			drug_name = form.cleaned_data['drug_name']
 			if form.cleaned_data['all_deleted']: # delete all reminder objects
 				ReminderTime.objects.filter(
 					to=patient, prescription__drug__name__iexact=drug_name, 
 					reminder_type=ReminderTime.REFILL).delete()
-			for r in form.cleaned_data['reminders_for_deletion']:
+			for r in form.cleaned_data['reminders_for_deletion']: 
 				r.delete()
 			return HttpResponse('')
 	return HttpResponseBadRequest('Something went wrong')
